requirements.txt
# requirements.txt
python-dotenv
pydantic
pandas
sqlalchemy
streamlit
openai
requests
numpy
groq